Changed element to observe to #queue-overlay
#music-content (big playlist or album or 'See All' song list) isn't always on screen. #queue-overlay is that mini pop-up que in the bottom right corner on the player toolbar.
But there is a problem: #queue-overlay is dynamically filled when longer than 10, meaning if it isn't visible and the song playing isn't in the first 10 in the que, there's no <tr class='current-playing'> to fetch rating from. If the filled #music-content is on screen at the moment, rating fill be fetched. If not then the user should open the minique and change the song or something.
You can thank Google for not providing 'data-rating' attribute anywhere when no currently playing playlist is open.